+++ smawg/__init__.py
@@ -3,13 +3,13 @@
 See https://github.com/expurple/smawg for more info about the project.
 """
 
-from smawg._engine import Game, Hooks, validate
+from smawg._engine import Game, Hooks, InvalidAssets, validate
 from smawg._plugin_interface import Ability, Combo, Player, Race, Region
-from smawg.exceptions import InvalidAssets, RulesViolation
+from smawg.exceptions import RulesViolation
 
 
 __all__ = [
-    "Game", "Hooks", "validate",
+    "Game", "Hooks", "InvalidAssets", "validate",
     "Ability", "Combo", "Player", "Race", "Region",
-    "InvalidAssets", "RulesViolation"
+    "RulesViolation"
 ]